refactor: 继续拆分独立窗口模式hooks和配置文件
